Add logic to update task list based on active project
There are two main changes in this commit. The first one is the added
functionality of having an activeProject variable in localStorage which
is updated by clicking the project names in the list. The second one is
the updated logic of loadTasks() so that it always loads the tasks in
of the current active project.